PageNav component, tests and scss
