+++ WhaleSpotting/react-app/src/components/PageNav.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+export default function PageNav() {
+    return (
+        <div className="page-nav" data-testid="page-nav">
+            <button className="page-button"> Previous Page</button>
+            <button className="page-button" >Next Page</button>
+        </div>
+    );
+}
+
